# shared/contracts/openapi/advertified-commercial-api.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Advertified Commercial API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tenants/{tenantId}/opportunities": {
      "get": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "ListOpportunities",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityViewCursorPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "CreateOpportunity",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key for safely retrying this command.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required for unsafe cookie-authenticated browser requests.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpportunityCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Strong optimistic-concurrency version for this entity.",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "True when a prior canonical command result was replayed.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenantId}/opportunities/{opportunityId}": {
      "get": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "GetOpportunity",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "opportunityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetailView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "UpdateOpportunity",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "opportunityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key for safely retrying this command.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Strong ETag version returned by the latest read.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required for unsafe cookie-authenticated browser requests.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOpportunityCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Strong optimistic-concurrency version for this entity.",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "True when a prior canonical command result was replayed.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenantId}/strategies/{strategyId}": {
      "get": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "GetStrategy",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "strategyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenantId}/agent-runs/{runId}": {
      "get": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "GetAgentRun",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenantId}/human-tasks": {
      "get": {
        "tags": [
          "Opportunity qualification"
        ],
        "operationId": "ListHumanTasks",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Optional UUID request correlation identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanTaskViewCursorPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "UUID support reference for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanSafeProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOpportunityCommand": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "nullable": true
          },
          "sourceRef": {
            "type": "string",
            "nullable": true
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedValueMinor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "problemSummary": {
            "type": "string",
            "nullable": true
          },
          "objectiveSummary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HumanSafeProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "HumanTaskViewCursorPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HumanTaskView"
            },
            "nullable": true
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OpportunityDetailView": {
        "type": "object",
        "properties": {
          "opportunity": {
            "$ref": "#/components/schemas/OpportunityView"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceSourceView"
            },
            "nullable": true
          },
          "evidenceItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceItemView"
            },
            "nullable": true
          },
          "evidenceSet": {
            "$ref": "#/components/schemas/EvidenceSetView"
          },
          "interpretation": {
            "$ref": "#/components/schemas/BusinessInterpretationView"
          },
          "angles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpportunityAngleView"
            },
            "nullable": true
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategyVersionView"
          },
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRunView"
            },
            "nullable": true
          },
          "nextAction": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OpportunityView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "nullable": true
          },
          "sourceRef": {
            "type": "string",
            "nullable": true
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "stage": {
            "type": "string",
            "nullable": true
          },
          "expectedValueMinor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "problemSummary": {
            "type": "string",
            "nullable": true
          },
          "objectiveSummary": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OpportunityViewCursorPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpportunityView"
            },
            "nullable": true
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}